admin/requirements.txt
Flask==2.1.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
Flask-Migrate==3.1.0
Werkzeug==2.3.6
flask-swagger-ui==4.11.1